Fix filtering of mutations based on associated genes when "--genes" option is provided.
Fix filtering of mutations based associated genes when "--genes"
option is provided.

It was accidentially broken in commit 6a479bd6772770c17a0de522cae584a6c8ab8de2:

    If a set of genes was provided, only keep those associated genes
    for the mutation that appear in this set of genes.

due a wrong if condition which caused that mutations that had
associated genes which where not in the set of genes, where not
removed.

Signed-off-by: Gert Hulselmans <gert.hulselmans@kuleuven.be>