Fix scoring with MotifLocator when only 1 or 2 different INCLUSive motif files with motifs are used.
Fix scoring with MotifLocator when only 1 or 2 different INCLUSive
motif files with motifs are used instead of all 3.

When iterating over a list and deleting the current element of the list,
the next iteration of the for loop skips an element.
So intead loop over a copy of the list.

Signed-off-by: Gert Hulselmans <gert.hulselmans@kuleuven.be>